Go 1.13: Fix go mod permissions for user 1000:1000
With cloudogu/golang:1.12.7-1 it was possible to compile with the unprivileged
user 1000, specifically to have `go mod` cache dependencies under
`/go/pkg/mod/`. With 1.13.10-1 this is no longer possible as the directories
are owned by root.

This commit chowns the /go directory recursively to the user 1000 to re-enable
compiling with `go mod`.